delete old Products Page files
