+++ src/containers/app/AppContainer.js
@@ -12,11 +12,9 @@ import AppHomePage from './AppHomePage';
 import CareProductPage from '../products/CareProductPage';
 import GalleryProductPage from '../platform/GalleryProductPage';
 import OrganizationsProductPage from '../platform/OrganizationsProductPage';
-import OtherProductsPage from '../products/OtherProductsPage';
 import OtherServicesPage from '../platform/OtherServicesPage';
 import PCMProductPage from '../products/PCMProductPage';
-import Products from '../products/NewProductsPage';
-import ProductsPage from '../products/ProductsPage';
+import Products from '../products/ProductsPage';
 import Platform from '../platform/Platform';
 import WorkingWithUs from '../workingwithus/WorkingWithUs';
 
@@ -27,9 +25,7 @@ const AppContainer = () => (
     <Route exact path={Routes.ORGANIZATIONS} component={OrganizationsProductPage} />
     <Route exact path={Routes.CARE} component={CareProductPage} />
     <Route exact path={Routes.PCM} component={PCMProductPage} />
-    <Route exact path={Routes.OTHER_PRODUCTS} component={OtherProductsPage} />
-    <Route exact path={Routes.PRODUCTS} component={ProductsPage} />
-    <Route exact path={Routes.PRODUCTS_NEW} component={Products} />
+    <Route exact path={Routes.PRODUCTS} component={Products} />
     <Route exact path={Routes.PLATFORM} component={Platform} />
     <Route exact path={Routes.WORKING_WITH_US} component={WorkingWithUs} />
     <Route exact path={Routes.ABOUT} component={About} />
