+++ src/containers/products/ProductsList.js
@@ -7,7 +7,7 @@ import styled from 'styled-components';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faChevronDown, faChevronUp } from '@fortawesome/pro-regular-svg-icons';
 
-import ProductTile from './NewProductTile';
+import ProductTile from './ProductTile';
 import ProductTileWithImage from './ProductTileWithImage';
 import PageSection from '../../components/layout/NewPageSection';
